core/requirements.txt
abnf==2.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
cachetools==6.2.4
cbor2==5.8.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
cint==1.0.0
claude-agent-sdk==0.1.18
claude-code-sdk==0.0.25
click==8.3.1
cryptography==46.0.3
cuid2==2.0.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
fastapi==0.128.0
fickling==0.1.6
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
gql==4.0.0
graphql-core==3.2.7
graphviz==0.21
grpclib==0.4.9
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
intervaltree==3.2.1
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaitaistruct==0.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.25.0
mdurl==0.1.2
mmh3==5.2.0
modal==1.3.0.post1
multidict==6.7.0
networkx==3.6.1
openai==2.14.0
packaging==25.0
pdfminer.six==20250506
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
polyfile-weave==0.5.7
postgrest==2.27.0
propcache==0.4.1
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pyroaring==1.0.3
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
realtime==2.27.0
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
ruff==0.14.10
sentry-sdk==2.48.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==3.1.1
starlette==0.50.0
stdlib-list==0.11.1
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
synchronicity==0.11.1
tenacity==9.1.2
toml==0.10.2
tqdm==4.67.1
typer==0.21.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
uvloop==0.22.1
wandb==0.23.1
watchfiles==1.1.1
weave==0.52.22
websockets==15.0.1
yarl==1.22.0
